backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.54
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0极速.6
email-validator==2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
